models.py: add __str__ methods for clear admin page
Until now, only a few models implemented the dunder str method which
returns a representetive string for them. When this method is not
present, a default 'XXX Object' would be returned. This contains no
information about the record, which makes admin tasks a nightmare.

With this commit, almost all models implement the dunder str method
which will explain their roots and connections. Now, the
administrative tasks are more clear and less error prone.